[project @ 1998-02-25 19:29:52 by sof] Dictionaries are now named as follows:
  $d<class><tycon><n>

where "n" is a positive int, "tycon" is the name of the tyvar/tycon
of the first argument to the "class" that the dict represent an
instance of.

The change should improve the behaviour of the recompilation
checker, preventing the recompilation of all the dependents
of a module whenever a data type of instance is added to it.
(The common behaviour should be no recompilations, but there
are cases where the naming scheme fails to prevent a recompile.)