[project @ 1998-02-25 19:48:54 by sof] Interface file version checking support. Added a compiler version field to the interface file header. The format is now
  _interface_ <IfaceName> <ModuleVersion> <CompilerVersion>

where the compiler version follow the value of $(ProjectVersionInt).
Any mismatch in version numbers causes the renamer to give up.

A compiler version number of 0 means turn off version checking (used
by PrelGHC.hi to avoid having to update every time we release.)

.hi-boot files are treated specially, the absence of a compiler
version number in the header is taken to mean that there was a `0'.
Need to do this since hsc's .hi-boot files have to also be useable
by versions of the compiler that don't grok version info in interface
files (e.g., ghc-2.10.)