[project @ 1999-01-13 17:25:37 by simonm] Added a generational garbage collector.
The collector is reliable but fairly untuned as yet.  It works with an
arbitrary number of generations: use +RTS -G<gens> to change the
number of generations used (default 2).

Stats: +RTS -Sstderr is quite useful, but to really see what's going
on compile the RTS with -DDEBUG and use +RTS -D32.

ARR_PTRS removed - it wasn't used anywhere.

Sanity checking improved:
	- free blocks are now spammed when sanity checking is turned on
	- a check for leaking blocks is performed after each GC.