[project @ 1999-10-05 10:30:26 by simonmar] Build System Cleanup --------------------
This commit is intended to clear the way for automatic RPM building.
In particular, the idea is that 'make install' doesn't build anything,
so you can do

	make install prefix=/some/tmp/prefix

to install everything in /some/tmp/prefix, but leave the hardwired-in
install directories the same.

For scripts that depend on the install dir (currently just the GHC
driver and mkdependHS), we now build two versions: <script>-inplace,
which is used for running the script from the build tree, and
<script>, which is the to-be-installed version.

NOTE: binary distributions are now a little bit trickier to build.
You *must* include the line "BIN_DIST=1" in your build.mk if you
intend to make a binary distribution from the current build tree.
This is because certain scripts have to be built differently, and we
don't rebuild them when doing 'make binary-dist' anymore (since 'make
binary-dist' just does a 'make install' with a re-targetted prefix,
just like the RPM builder).

Other changes

	- the binary-dist machinery is now all in fptools/Makefile
	- removed a gratuitous $(package)-$(version) level of
	  directories from the binary distribution.
	- binary distributions are now placed under the package
	  name, rather than fptools/fptools.
	- various other minor cleanups.