[project @ 1999-10-19 11:01:24 by sewardj] Reimplement back-end for foreign import (calling out).
Return to a cleaned-up version of Alastair's callfun.S,
wherein an architecture and calling-convention specific
piece of assembly code is used to construct arguments and
then call the specified function, under the direction of
a type descriptor string.

Defined an interface to this function (universal_call_c)
which I hope will work regardless of 32-or-64 bitness,
endianness and calling convention.

Current implementation is for x86-linux only.