[project @ 1999-11-29 17:34:14 by simonpj] Make it so that a class decl generates default method decls for every method, not just for the ones that the user supplies default-methods for.
GHC will never call these default-default methods, because
when it finds an instance decl with no defn for a method,
*and* the class decl doesn't have a user-programmed default
method, it whips up a new default method for that instance
decl so that the error message is more informative than
the default-default method would be.

But Hugs isn't so smart, and wants to call something from
the class decl.

This change required fiddling with more than I expected.  Sigh.

Simon