[project @ 2000-01-22 18:00:03 by simonmar] Fix bug in async exception handling: the target TSO may have been relocated as a result of a stack overflow.
Introduce a new StgTSOWhatNext value "ThreadRelocated", which
indicates that this TSO has moved, and the new location is in the link
field.  The garbage collector shorts these out just like indirections.

We have to check for relocated TSOs in killThread# (and any other
primops which take a ThreadId# as an argument - there aren't any at
present).