[project @ 2000-02-04 11:15:04 by simonmar] Fix bug #1 in the unregisterised RTS:
Now that catch# is a fully-fledged IO operation, it has to respect the
return convention for IO, in particular if we don't have an R1
(eg. when unregisterised) then the return value goes on the stack.

The return convention could be better, after all we use R1 when
returning normal values, but I'll just fix it for now.