[project @ 2000-07-24 15:16:44 by simonmar] Sigh, change the package definition again.  We weren't making a distinction between libraries which need a suffix tag (eg. libHSstd_p), and those that don't (eg. libHSstd_cbits).
The package spec now has two components for libraries, hs_libraries
(tagged) and extra_libraries (untagged).  The ordering of these
components is important: we specify that hs_libraries are linked
before extra_libraries.  This sounds problematic if you want to link
some plain C libraries before some Haskell libs, but in these cases it
should be possible to create separate packages for the two libs and
specify the dependencies explicitly.